control fragment buffer size
if audio and video are read from the same datasource then the audio samples will arrive with much higher rate than they can be written as video needs to keep up with audio. As buffering is done in heap it might lead to OOME. There we restrict the buffer to 10 fragments of each. Once the limit is reached the source track's thread will be blocked